Exception syntax for file existence was added.

+++ ubuntu/18.04/amp.sh
@@ -43,14 +43,18 @@ done
 
 hostnamectl set-hostname "$HOSTNAME"
 
-printf "\n\nSetting up hosts ... \n"
-cp /etc/hosts /etc/hosts.bak
-sed -i "1 a\127.0.1.1 $HOSTNAME" /etc/hosts
+if [ -f /etc/hosts ]; then
+  printf "\n\nSetting up hosts ... \n"
+  cp /etc/hosts /etc/hosts.bak
+  sed -i "1 a\127.0.1.1 $HOSTNAME" /etc/hosts
+fi
 
 # This will cause the set+update hostname module to not operate (if true)
-printf "\n\nSetting up cloud.cfg ... \n"
-cp /etc/cloud/cloud.cfg /etc/cloud/cloud.cfg.bak
-sed -i -E -e '/preserve\_hostname\s{0,}?\:/{ s/\:.*/\: true/; }' /etc/cloud/cloud.cfg
+if [ -f /etc/cloud/cloud.cfg ]; then
+  printf "\n\nSetting up cloud.cfg ... \n"
+  cp /etc/cloud/cloud.cfg /etc/cloud/cloud.cfg.bak
+  sed -i -E -e '/preserve\_hostname\s{0,}?\:/{ s/\:.*/\: true/; }' /etc/cloud/cloud.cfg
+fi
 
 
 #
